# RDL.Platform.API/api/openapi.yaml
openapi: 3.1.0
info:
  title: Platform API
  version: 0.7.0
  description: |
    Identidad y tenancy del SaaS de facturación electrónica.
    La organización activa sale SOLO del claim `org_id` del JWT verificado (emitido por el hook de Supabase Auth)
    y de una membresía activa revalidada en cada request. Ningún `organization_id` en body, query, headers o ruta
    cambia el tenant.
    TODO(contracts): alinear con el OpenAPI de Platform del repo de contratos cuando exista.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /v1/organizations/current/invitations:
    post:
      summary: Invita a una persona por email con un rol (owner, admin)
      description: |
        Exige Idempotency-Key. El token en claro solo viene en la primera respuesta (la base guarda su hash);
        un reintento con la misma clave devuelve la misma invitación sin token (ADR 0006).
        Solo un owner puede invitar con rol owner. 409 si ya hay una pendiente para ese email o ya es miembro activo.
        TODO(notificaciones): hoy el enlace no se envía por email.
      operationId: createInvitation
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201":
          description: Invitación creada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    get:
      summary: Invitaciones de la organización activa (owner, admin)
      operationId: listInvitations
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, accepted, revoked, expired] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: Página de invitaciones
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Invitation" } }
                  nextCursor: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/organizations/current/invitations/{id}:
    delete:
      summary: Revoca una invitación (owner, admin). Idempotente.
      operationId: revokeInvitation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Revocada (o ya estaba revocada o vencida) }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /v1/invitations/{token}/accept:
    post:
      summary: Acepta una invitación y crea la membresía
      description: |
        El email del usuario autenticado debe coincidir con el de la invitación (si no, 404: no se revela nada).
        El usuario puede pertenecer ya a otras organizaciones (caso contador). Si no tenía organización activa,
        esta queda activa. Refresque la sesión (y elija la organización si hace falta) para operar en ella.
      operationId: acceptInvitation
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Aceptada
          content:
            application/json:
              schema:
                type: object
                required: [organizationId, role, activeOrganizationSelected, tokenRefreshRequired]
                properties:
                  organizationId: { type: string, format: uuid }
                  role: { $ref: "#/components/schemas/Role" }
                  activeOrganizationSelected: { type: boolean }
                  tokenRefreshRequired: { type: boolean, const: true }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "410": { $ref: "#/components/responses/Problem" }
  /v1/organizations/current/branches:
    get:
      summary: Sucursales de la organización activa (todos los roles)
      operationId: listBranches
      parameters:
        - { name: active, in: query, schema: { type: boolean } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: Página de sucursales
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Branch" } }
                  nextCursor: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    post:
      summary: Crea una sucursal (owner, admin). Idempotente.
      operationId: createBranch
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code, name]
              properties:
                code: { type: string, pattern: "^[A-Za-z0-9_-]{1,20}$", description: "Único por organización e inmutable." }
                name: { type: string, maxLength: 150 }
                address: { type: string, maxLength: 500 }
                phone: { type: string, maxLength: 30 }
                email: { type: string, format: email }
      responses:
        "201":
          description: Creada (o repetición idempotente)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/organizations/current/branches/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Detalle de una sucursal (todos los roles)
      operationId: getBranch
      responses:
        "200":
          description: Sucursal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      summary: Edita o desactiva una sucursal (owner, admin)
      description: Campo ausente = sin cambio; "" en address, phone o email lo borra; isActive=false es la baja lógica. El código no se edita.
      operationId: updateBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, maxLength: 150 }
                address: { type: string, maxLength: 500 }
                phone: { type: string, maxLength: 30 }
                email: { type: string }
                isActive: { type: boolean }
      responses:
        "200":
          description: Actualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT (Supabase Auth, ES256, verificado con JWKS)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 1 a 255 caracteres ASCII visibles. Se recuerda 24 h.
      schema: { type: string, minLength: 1, maxLength: 255 }
...
  responses:
    Unauthenticated:
      description: Falta el token o no es válido
      headers:
        WWW-Authenticate: { schema: { type: string } }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: |
            Estable por error: urn:rdl:platform:problem:{slug}. Slugs actuales: unauthenticated,
            no-active-organization, membership-inactive, forbidden, not-found, method-not-allowed,
            malformed-request, validation, conflict, email-required, user-disabled, idempotency-key-required,
            idempotency-key-reused, last-owner, owner-required, invitation-expired, invitation-not-pending,
            internal, unavailable.
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    Invitation:
      type: object
      required: [id, email, role, status, expiresAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        status: { type: string, enum: [pending, accepted, revoked, expired] }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        token: { type: string, description: "Solo en la primera respuesta del alta." }
        acceptPath: { type: string, description: "Solo en la primera respuesta del alta." }
    Branch:
      type: object
      required: [id, code, name, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Role:
      type: string
      enum: [owner, admin, biller, collector, accountant, read_only]
...